Revert "move shared components into /shared folder"
This reverts commit 56c2c6daf668dea5dced3854e35c4de3711f3ae0.
