+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-id].page/entity-page-wrapper/entity-page-header.tsx
@@ -6,8 +6,8 @@ import { useRouter } from "next/router";
 import slugify from "slugify";
 import { generateEntityLabel } from "../../../../../lib/entities";
 import { TopContextBar } from "../../../../shared/top-context-bar";
-import { OntologyChip } from "../../../shared/ontology-chip";
-import { HashOntologyIcon } from "../../../shared/hash-ontology-icon";
+import { OntologyChip } from "../../../types/entity-type/ontology-chip";
+import { PlaceholderIcon } from "../../../types/entity-type/placeholder-icon";
 import { useEntityEditor } from "../entity-editor-context";
 
 export const EntityPageHeader = () => {
@@ -48,7 +48,7 @@ export const EntityPageHeader = () => {
       <Box py={3.75}>
         <Container>
           <OntologyChip
-            icon={<HashOntologyIcon />}
+            icon={<PlaceholderIcon />}
             domain="hash.ai"
             path={
               <>
