+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/[entity-type-id].page.tsx
@@ -10,8 +10,8 @@ import { useInitTypeSystem } from "../../../../lib/use-init-type-system";
 import { getPlainLayout, NextPageWithLayout } from "../../../../shared/layout";
 import { TopContextBar } from "../../../shared/top-context-bar";
 import { EditBar } from "./edit-bar";
-import { HashOntologyIcon } from "../../shared/hash-ontology-icon";
-import { OntologyChip } from "../../shared/ontology-chip";
+import { HashOntologyIcon } from "./hash-ontology-icon";
+import { OntologyChip } from "./ontology-chip";
 import { PropertyListCard } from "./property-list-card";
 import {
   PropertyTypesContext,
