+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/empty-property-list-card.tsx
@@ -9,7 +9,7 @@ import {
   Theme,
   Typography,
 } from "@mui/material";
-import { WhiteCard } from "../../shared/white-card";
+import { WhiteCard } from "./white-card";
 
 const cardActionHoverBlue: SxProps<Theme> = (theme) => ({
   [`.${cardActionAreaClasses.root}:hover &`]: {
