+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/hash-ontology-icon.tsx
