+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/ontology-chip.tsx
