+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/property-menu.tsx
@@ -25,10 +25,7 @@ import {
   usePopupState,
 } from "material-ui-popup-state/hooks";
 import { Fragment, useId } from "react";
-import {
-  OntologyChip,
-  parseUriForOntologyChip,
-} from "../../shared/ontology-chip";
+import { OntologyChip, parseUriForOntologyChip } from "./ontology-chip";
 
 const parseVersion = (id: string) => {
   const uri = validateVersionedUri(id);
