+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/white-card.tsx
